Changed impl. to pass ExecImplOptions into exec functions.
This is a much more clean design.
But some functionailty is temporary disabled due to Stdio is
completly opaqueue.i

Now all options which are relevant for execution are stored in
ExecImplOptions inside of Command and only the trait objects
and the expected exit status are directly stored inside of
Command and neither is needed for the command spawning.